adk/adk_project_01/pyproject.toml
[project]
name = "adk-project-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[extensions]>=1.28.0",
]
